requirements.txt
# Core ML/AI libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
# bitsandbytes is GPU-only, skip on Mac CPU
# bitsandbytes>=0.41.0
trl>=0.7.0

# Unsloth is optional - only needed if fine-tuning locally
# For Mac/CPU, use a service provider (Colab, Kaggle, etc.) instead
# If you have GPU and want to install locally:
#   pip install unsloth>=2024.1
# Note: unsloth[colab-new] includes xformers which doesn't build on Mac

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Sentiment analysis
vaderSentiment>=3.3.2
textblob>=0.17.1

# Gender detection
gender-guesser>=0.4.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
huggingface-hub>=0.19.0
requests>=2.31.0

# Optional: for Task 3 augmentation
# anthropic>=0.18.0  # For Claude API
# google-generativeai>=0.3.0  # For direct Google Gemini API (recommended)
# openrouter>=1.0.0  # For Grok via OpenRouter
# Uncomment if using API for augmentation

# Optional: for better text processing
nltk>=3.8.1
spacy>=3.7.0

